% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{boxplot_label_outliers}
\alias{boxplot_label_outliers}
\title{Standard boxplot with labeled outliers}
\usage{
boxplot_label_outliers(M, ...)
}
\arguments{
\item{M}{one-dimensional matrix of values with row names for labels}

\item{...}{standard commands for \code{boxplot()} function}
}
\value{
standard output of \code{boxplot()} function
}
\description{
Any point outside of the whiskers will be labeled. Passes all standard 
\code{boxplot()} parameters through to the R \code{boxplot()} function.
}
\examples{
M = matrix( c(3,4,7,5,6,10), nrow=6, ncol=1 )
dimnames(M)[[1]] = paste("drug",1:6)
boxplot_label_outliers(M[,1])
}

