% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy.R
\name{chou_synergy}
\alias{chou_synergy}
\title{Calculate statistics needed for Chou Synergy
Curve fitting is performed by the \code{drm()} function in the \code{drc} 
library. To fit the curve, you need to select a non-linear function. To 
estimate the slope, upper asymptote, lower asymptote, and EC50, pass 
drc::LL.4(). To fix the lower asymptote at 1 and estimate the other 
parameters, pass drc::LL.3(). To fix the upper asympotote at 1 and the lower 
asymptote at 0, pass dcr::LL.2. For a list of available functions, see 
\code{drc::getMeanFunctions()}.}
\usage{
chou_synergy(ds, sample_type, treatment_1, treatment_2, hour, proportion_1, fct,
  summary_method)
}
\arguments{
\item{ds}{dataset}

\item{sample_type}{sample type in ds}

\item{treatment_1}{treatment in ds}

\item{treatment_2}{treatment in ds}

\item{hour}{hour in ds}

\item{proportion_1}{value between 0 and 1, indicating the fixed proportional
relationship between concentrations of treatment_1 and treatment_2 in the 
combined observations. If 50/50, pass 0.5.}

\item{fct}{Non-linear function to fit, e.g. drc::LL.3(). See summary.}

\item{summary_method}{mean or median}
}
\value{
A list with the Chou Synergy helper values for treatment_1, 
treatment_2, the combination of treatment_12, and the combination index 
data frame (itself containing columns for Fa, Fu, and CI).
}
\description{
Ting-Chao Chou Pharmacological Reviews 2006
}

