% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy.R
\name{chou_synergy_CI_median}
\alias{chou_synergy_CI_median}
\title{Plot single-value Chou Combination Index at IC50}
\usage{
chou_synergy_CI_median(CS, proportion_1)
}
\arguments{
\item{CS}{chou statistics calculated by \code{\link{chou_synergy}}}

\item{proportion_1}{proportion of dose 1 vs. dose 2, numeric between 0 and 1}
}
\value{
Combination Index
}
\description{
Plot single-value Chou Combination Index at IC50
}

