% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{convert_to_foldchange}
\alias{convert_to_foldchange}
\title{Convert values in M1, M2 to fold-change M1 vs M2}
\usage{
convert_to_foldchange(M1, M2)
}
\arguments{
\item{M1}{matrix for numerator of fold-change}

\item{M2}{matrix for denominator of fold-change}
}
\value{
matrix with fold-change for M1/M2. Values that are NA in either M1 
or M2 are NA in M
}
\description{
Convert values in M1, M2 to fold-change M1 vs M2
}
\examples{
M1 = matrix( c(1, 1, 4, 1), nrow=2, ncol=2 )
M2 = matrix( 1:4, nrow=2, ncol=2 )
convert_to_foldchange(M1, M2)
}

