% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{create_dataset}
\alias{create_dataset}
\title{Create a dataset from raw data without plate}
\usage{
create_dataset(sample_types, treatments, concentrations, hours, values,
  plate_id, negative_control = NA)
}
\arguments{
\item{sample_types}{vector of sample types}

\item{treatments}{vector of treatments}

\item{concentrations}{vector of concentrations}

\item{hours}{vector of timepoints}

\item{values}{vector of measured response to treatment}

\item{plate_id}{experiment identification string, useful if multiple datasets 
are later combined.}

\item{negative_control}{If an empty string, there are no negative control 
wells. If a number such as 0, each treatment will be expected to contain 
wells with this concentration as a negative control. If a string, the 
name of treatment wells which contain the universal negative control for this 
plate. If a data frame, a mapping of each vehicle to each drug individually.}
}
\value{
a data frame where columns indicate the sample type, treatment, 
concentration, observed raw value, normalized value (raw until normalization 
is run), name of the negative_control treatment, whether a particular row 
is a negative control for at least one other row, hours since the start time,
and plate of origin
}
\description{
Create a dataset from raw data without plate
}
\examples{
# six measurements: DMSO, 100, and 200 nM for two drugs. 
# plan to normalize each line against DMSO for that line
ds = create_dataset( 
  sample_types= c("line1","line1","line1","line2","line2","line2"),
  treatments = c("DMSO","drug1","drug2","DMSO","drug1","drug2"),
  concentrations = c(0, 100, 200, 0, 100, 200),
  hours = c(48, 48, 48, 48, 48, 48),
  values = c(98, 90, 20, 99, 89, 87), 
  plate_id = "plate_1",
  negative_control = "DMSO")
  
# six measurements; drug1 at 0, 100, 200 nM and drug2 at 0, 100, 200 nM. 
# plan to normalize against zero concentration for each line
ds = create_dataset( 
  sample_types= c("line1","line1","line1","line2","line2","line2"),
  treatments = c("drug1","drug1","drug2","drug2","drug2","drug2"),
  concentrations = c(0, 100, 200, 0, 100, 200),
  hours = c(48, 48, 48, 48, 48, 48),
  values = c(98, 90, 20, 99, 89, 87), 
  plate_id = "plate_1",
  negative_control = 0)
  
# six measurements; drug1 at 0, 100, 200 nM and drug2 at 0, 100, 200 nM. 
# plan to normalize drug1 against DMSO and drug2 against ethanol
individual_vehicles = data.frame(
  drug=c("drug1", "drug2"), 
  vehicle=c("DMSO", "ethanol"),
  stringsAsFactors=FALSE)
ds = create_dataset( 
  sample_types= c("line1","line1","line1","line2","line2","line2"),
  treatments = c("DMSO","drug1","drug1","ethanol","drug2","drug2"),
  concentrations = c(0, 100, 200, 0, 100, 200),
  hours = c(48, 48, 48, 48, 48, 48),
  values = c(98, 90, 20, 99, 89, 87), 
  plate_id = "plate_1",
  negative_control = individual_vehicles)
}

