% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{create_empty_plate}
\alias{create_empty_plate}
\title{Create a plate list with empty elements}
\usage{
create_empty_plate(number_of_wells, hour = 0, plate_id)
}
\arguments{
\item{number_of_wells}{plate size, one of {6, 24, 96, 384}}

\item{hour}{hours since the experiment began}

\item{plate_id}{string that identifies this plate map}
}
\value{
a data frame
}
\description{
Create a plate list with empty elements
}
\examples{
create_empty_plate( 96, hour=0, plate_id="plate_1")
create_empty_plate( 384, hour=12, plate_id="plate_9")
}
\seealso{
\code{\link{create_empty_plate_map}}
}

