% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{create_empty_plate_map}
\alias{create_empty_plate_map}
\title{Create an empty plate map}
\usage{
create_empty_plate_map(number_of_wells)
}
\arguments{
\item{number_of_wells}{plate size, one of {6, 24, 96, 384}}
}
\value{
a plate_map list with matrixes for treatment, concentration, 
sample_types, density, and passage
}
\description{
The plate map object is a list of identically sized matrixes named 
treatment, concentration, sample_type, density, and passage.
}
\examples{
create_empty_plate_map( 96 )
}
\seealso{
\code{\link{create_empty_plate}}
}

