% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{fit_DRC}
\alias{fit_DRC}
\title{Attempt to fit a dose-response curve from the dataset and calculate AUC.}
\usage{
fit_DRC(D, sample_types, treatments, hour, fct)
}
\arguments{
\item{D}{experiment dataset with columns matching the output of 
\code{combine_data_and_map}}

\item{sample_types}{sample types (e.g. distinct cell lines) to fit}

\item{treatments}{treatments to fit}

\item{hour}{hour at which to fit}

\item{fct}{Non-linear function to fit, e.g. drc::LL.3(). See summary.}
}
\value{
A HT_fit object
}
\description{
Given a data frame of measurements generated by 
\code{\link{combine_data_and_map}} or matching the format generated by that 
function, fit a dose-response curve for each unique sample_type/treatment 
specified by the sample_types and treatments parameters. A curve will 
be fit for each sample in sample_types, at each treatment in treatments
}
\section{Non-linear function for curve fitting}{


Curve fitting is performed by the \code{drm()} function in the \code{drc} 
library. To fit the curve, you need to select a non-linear function. To 
estimate the slope, upper asymptote, lower asymptote, and EC50, pass 
drc::LL.4(). To fix the lower asymptote at 1 and estimate the other 
parameters, pass drc::LL.3(). To fix the upper asympotote at 1 and the lower 
asymptote at 0, pass dcr::LL.2. For a list of available functions, see 
\code{drc::getMeanFunctions()}. 

To call this function you must load the drc package in your R session.
}
\examples{
sample_types = rep( c(rep("line1",3), rep("line2",3)), 5)
treatments = c(rep("DMSO",6), rep("drug",24))
concentrations = c( rep(0,6),rep(200,6), rep(500,6),rep(1000,6),rep(5000,6))
values=c(100,99,100,90,91,92,99,97,99,89,87,88,86,89,88,56,59,58,66,65,67,
         25,23,24,42,43,46,4,5,9)
hours = rep(48, length(values))
plate_id = "plate_1"
ds = create_dataset( sample_types, treatments, concentrations, 
                      hours, values, plate_id, negative_control = "DMSO")
ds = normalize_plates_by_vehicle(ds, summary_method = "mean")
library(drc)
# Fit model using three-parameter log-logistic function
fit_DRC(ds, sample_types=c("line1", "line2"), treatments=c("drug"), 
        hour = 48, fct=drc::LL.3() )
 
# Fit model using four-parameter log-logistic function
fit_DRC(ds, sample_types=c("line1", "line2"), treatments=c("drug"), 
        hour = 48, fct=drc::LL.4() )
}

