% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{get_concentrations}
\alias{get_concentrations}
\title{Sorted unique list of treatment concentrations}
\usage{
get_concentrations(D)
}
\arguments{
\item{D}{A data frame matching the format of one generated by 
\code{combine_data_and_map}}
}
\value{
sorted vector of concentrations, including 0 if present
}
\description{
Convenience method to extract treatment concentrations
}
\examples{
ds = create_dataset( 
  sample_types= c("line1","line1","line1","line2","line2","line2"),
  treatments = c("DMSO","drug1","drug2","DMSO","drug1","drug2"),
  concentrations = c(0, 100, 200, 0, 100, 200),
  hours = c(48, 48, 48, 48, 48, 48),
  values = c(100, 90, 20, 100, 89, 87), 
  plate_id = "plate_1",
  negative_control = "DMSO")
get_concentrations( ds )
}
\seealso{
\code{\link{get_sample_types}}

\code{\link{get_treatments}}

\code{\link{get_hours}}

\code{\link{combine_data_and_map}}
}

