% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{get_hours}
\alias{get_hours}
\title{Sorted unique list of hours since time zero}
\usage{
get_hours(D)
}
\arguments{
\item{D}{A dataframe generated by \code{combine_data_and_map}}
}
\value{
sorted vector of hours
}
\description{
Convenience method to extract a sorted list of the hours
}
\examples{
ds = create_dataset( 
  sample_types= c("line1","line1","line1","line2","line2","line2"),
  treatments = c("DMSO","drug1","drug2","DMSO","drug1","drug2"),
  concentrations = c(0, 100, 200, 0, 100, 200),
  hours = c(48, 48, 48, 48, 48, 48),
  values = c(100, 90, 20, 100, 89, 87), 
  plate_id = "plate_1",
  negative_control = "DMSO")
get_hours( ds )
}
\seealso{
\code{\link{get_sample_types}}

\code{\link{get_treatments}}

\code{\link{get_hours}}
}

