% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{get_sample_types}
\alias{get_sample_types}
\title{Sorted unique list of sample types}
\usage{
get_sample_types(D)
}
\arguments{
\item{D}{A plate map list or a data frame generated by 
\code{combine_data_and_map}}
}
\value{
sorted vector of sample types
}
\description{
Convenience method to extract sample types. get_sample_types() can be called 
on either a plate map list or a data frame generated by 
\code{combine_data_and_map}.
}
\examples{
pm = create_empty_plate_map( 6 )
pm$sample_type[1,1:3] = "line1"
pm$sample_type[2,1:3] = "line2"
get_sample_types( pm )
ds = create_dataset( 
  sample_types= c("line1","line1","line1","line2","line2","line2"),
  treatments = c("DMSO","drug1","drug2","DMSO","drug1","drug2"),
  concentrations = c(0, 100, 200, 0, 100, 200),
  hours = c(48, 48, 48, 48, 48, 48),
  values = c(100, 90, 20, 100, 89, 87), 
  plate_id = "plate_1",
  negative_control = "DMSO")
get_sample_types( ds )
}
\seealso{
\code{\link{get_treatments}}

\code{\link{get_concentrations}}

\code{\link{get_hours}}
}

