% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{get_treatments}
\alias{get_treatments}
\title{Sorted unique list of treatments}
\usage{
get_treatments(D)
}
\arguments{
\item{D}{A plate map list or a data frame generated by 
\code{combine_data_and_map}}
}
\value{
sorted vector of treatments
}
\description{
Convenience method to extract treatments
}
\examples{
pm = create_empty_plate_map( 6 )
pm$treatment[1,1:3] = c("DMSO","drug1","drug2")
pm$treatment[2,1:3] = c("DMSO","drug1","drug2")
get_treatments( pm )
ds = create_dataset( 
  sample_types= c("line1","line1","line1","line2","line2","line2"),
  treatments = c("DMSO","drug1","drug2","DMSO","drug1","drug2"),
  concentrations = c(0, 100, 200, 0, 100, 200),
  hours = c(48, 48, 48, 48, 48, 48),
  values = c(100, 90, 20, 100, 89, 87), 
  plate_id = "plate_1",
  negative_control = "DMSO")
get_treatments( ds )
}
\seealso{
\code{\link{get_sample_types}}

\code{\link{get_concentrations}}

\code{\link{get_hours}}
}

