% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{metric_to_grid}
\alias{metric_to_grid}
\title{reshape a set of values to a matrix with axes defined by X any Y vectors}
\usage{
metric_to_grid(x_axis_labels, y_axis_labels, values)
}
\arguments{
\item{x_axis_labels}{vector of values that define the X axis}

\item{y_axis_labels}{vector of values that define the Y axis}

\item{values}{value to position at (X,Y) defined by the labels}
}
\value{
Matrix with columns named for x_axis_labels, rows named for 
y_axis_labels, and elements corresponding to the values parameter. X and Y 
labels will be sorted.
}
\description{
Useful for extracting two columns out of a data frame that will serve as 
column and row names, with a third column that will represent the value at 
that (row, column). The pairs identified by {x_axis_labels,y_axis_labels} 
should be unique; if not, the last value seen will be used.
}
\examples{
xlab = c("a","a","a","b","b","b","c","c","c")
ylab = c("x","y","z","x","y","z","x","y","z")
vals = c(1, 2 , 3 , 4 , 5 , 6 , 7 , 8, 9)
metric_to_grid(xlab, ylab, vals)
}

