% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{normalize_plates_by_vehicle}
\alias{normalize_plates_by_vehicle}
\title{Normalize raw data on each plate against the vehicle control wells}
\usage{
normalize_plates_by_vehicle(D, summary_method)
}
\arguments{
\item{D}{experiment dataset with columns matching the output of 
\code{\link{combine_data_and_map}}}

\item{summary_method}{string indicating how to summarize vehicle replicates, 
one of "mean" or "median"}
}
\value{
a data frame with the same columns and dimenson as D
}
\description{
Given a data frame of measurements generated by 
\code{\link{combine_data_and_map}}
or matching that format, divide each raw measurement on the specified plate 
by the summarized value for the appropriate negative control measurement on 
the same plate. Summary can be either mean or median. If the plate_id 
parameter is an empty string (default), all plates will be normalized.
}
\details{
If no normalization has been specified for D, the value_normalized column 
will be identical to the value column.
}
\examples{
# Create a six well plate testing two drugs on two cell lines. Each drug has 
# three concentrations, one of which is vehicle-only and labeled zero.
plate = create_empty_plate( 6, hour=0, plate_id="plate_1")
plate[1:2,1:3] = c(99,98,90,87,77,20)
plate_map = create_empty_plate_map( number_of_wells = 6 )
plate_map$concentration[1:2,1:3] = c( 0, 0, 100, 100, 200, 200 )
plate_map$treatment[1:2,1:3] = c("drug1","drug2","drug1","drug2","drug1",
                                 "drug2")
plate_map$sample_type[1:2,1:3] = rep( c("line1", "line2"), 3)
ds = combine_data_and_map( plate, plate_map, negative_control=0 )
normalize_plates_by_vehicle( ds, summary_method="mean")
}

