% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.HT_fit}
\alias{plot.HT_fit}
\title{Plot a dose response curve}
\usage{
\method{plot}{HT_fit}(x, ..., log10_xmin = 1, log10_xmax = 5,
  bar_multiple = 2, summary_method = "mean", show_EC50 = TRUE,
  show_legend = TRUE, show_x_log_tics = TRUE, show_x_exponent = TRUE)
}
\arguments{
\item{x}{HTfit object}

\item{...}{standard parameters for \code{plot} function}

\item{log10_xmin}{x minimum for plot on log10 scale, default 0}

\item{log10_xmax}{x maximum for plot on log10 scale, default 5}

\item{bar_multiple}{multiplier for standard error bars, default 2}

\item{summary_method}{summary method for points to plot in timecourse, one 
of ("mean", "median"), defaults "mean"}

\item{show_EC50}{show dotted vertical lines at EC50 values if they are fit,
default TRUE}

\item{show_legend}{show a legend, default TRUE}

\item{show_x_log_tics}{if axes==TRUE and this parameter is TRUE, draw tics 
between each power of 10 scaled appropriately. If axes==TRUE and this 
parameter is FALSE, omit ticks between powers of 10. Default TRUE.}

\item{show_x_exponent}{If axes==TRUE and this parameter is TRUE, show labels 
on X axis as 10^2, 10^3, ... . If axes==TRUE and this parameter is FALSE, 
show labels on X axis as an exponent (2, 3, ... ). Default TRUE.}
}
\value{
none
}
\description{
If the curve cannot be fit (meaning that the optimization method used by 
the drm() method from the drc package failes to converge) then the summarized 
points will be plotted without a curve connecting them.
}
\examples{
sample_types = rep( c(rep("line1",3), rep("line2",3)), 5)
treatments = c(rep("DMSO",6), rep("drug",24))
concentrations = c( rep(0,6),rep(200,6), rep(500,6),rep(1000,6),rep(5000,6))
values=c(100,99,100,90,91,92,99,97,99,89,87,88,86,89,88,56,59,58,66,65,67,
         25,23,24,42,43,46,4,5,9)
hours = rep(48, length(values))
plate_id = "plate_1"
ds = create_dataset( sample_types, treatments, concentrations, 
                      hours, values, plate_id, negative_control = "DMSO")
ds = normalize_plates_by_vehicle(ds, summary_method = "mean")
library(drc)
# Fit model using three-parameter log-logistic function
fit_1=fit_DRC(ds, sample_types=c("line1", "line2"), treatments=c("drug"), 
        hour = 48, fct=drc::LL.3() )
plot(fit_1)
plot(fit_1, show_EC50=FALSE, show_legend=FALSE, lwd=3,col=c("black", "gold"),
     log10_xmin=0, log10_xmax=4, xlab="concentration nM", 
     ylab="surviving fraction")
legend(1.5, 0.3, c("Line 1", "Line 2"), col=c("black", "gold"), pch=15)
}

