% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_chou_synergy_Fa_CI}
\alias{plot_chou_synergy_Fa_CI}
\title{plot FA vs. CI Chou synergy plot}
\usage{
plot_chou_synergy_Fa_CI(CS, ..., show_horizontal = TRUE)
}
\arguments{
\item{CS}{chou statistics}

\item{...}{standard parameters for \code{plot} function}

\item{show_horizontal}{show horizontal stripes}
}
\description{
plot FA vs. CI Chou synergy plot
}
\examples{
dose_SCH=c(0.1, 0.5, 1, 2, 4)
eff_SCH = c(0.6701, 0.6289, 0.5577, 0.4550, 0.3755)
dose_4HPR = c(0.1, 0.5, 1, 2)
eff_4HPR = c(0.7666, 0.5833, 0.5706, 0.4934)
eff_comb = c(0.6539, 0.4919, 0.3551, 0.2341)
syn = data.frame( 
    treatment_1 = rep("SCH66336", 13),
    conc_1 = c( dose_SCH, rep(0, 4), dose_SCH[1:4]),
    treatment_2 = rep("4-HPR", 13),
    conc_2 = c( rep(0, 5), dose_4HPR, dose_4HPR ),
    values = c(eff_SCH, eff_4HPR, eff_comb ) )
ds_lk = create_synergy_dataset(
                                sample_types = rep("sample_1", 13), 
                                treatments_1 = syn$treatment_1,
                                treatments_2 = syn$treatment_2,
                                concentrations_1 = syn$conc_1,
                                concentrations_2 = syn$conc_2,
                                values = syn$values)
CS_lk = chou_synergy( ds_lk, sample_type = "sample_1", hour=0,
                      treatment_1 = "SCH66336", treatment_2="4-HPR",
                      proportion_1=0.5,
                      fct=drc::LL.2(), summary_method="mean" )
plot_chou_synergy_Fa_CI(CS_lk)
}

