% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_color_grid}
\alias{plot_color_grid}
\title{convert a matrix to colors and plot a heatmap grid}
\usage{
plot_color_grid(M, block.height = 20, block.width = 10, space.X = 3,
  space.Y = 10, cex.x = 1, cex.y = 1, border = TRUE,
  color_palatte = c("blue", "white", "red"), color_bounds = NA,
  num_colors = 50)
}
\arguments{
\item{M}{matrix of values to plot}

\item{block.height}{element height, default 20}

\item{block.width}{element width, default 10}

\item{space.X}{space between elements on horizontal axis, default 3}

\item{space.Y}{space between elements on vertical axis, default 10}

\item{cex.x}{cex value for x axis}

\item{cex.y}{cex value for y axis}

\item{border}{boolean, draw a border around grid elements}

\item{color_palatte}{vector of colors c(cold,middle,hot), defaults to 
c("blue", "white", "red")}

\item{color_bounds}{two element vector with lower and upper limit for color 
palatte, defaults to c( min(M), max(M) )}

\item{num_colors}{number of color gradations in palatte, default 50}
}
\value{
M, including possible truncation from color bounds
}
\description{
If color_bounds
}
\examples{
M = matrix(1:12, nrow=3, ncol=4,byrow=TRUE)
plot_color_grid(M)
dimnames(M)[[1]] = c("row_a","row_b","row_c")
dimnames(M)[[2]] = c("col1", "col2", "col3","col4")
plot_color_grid(M, color_palatte=c("blue","white","yellow" ) )
}

