% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_synergy_interaction_index}
\alias{plot_synergy_interaction_index}
\title{Plot interaction index with confidence intervals for observed effects 
at combination doses having observed effects.}
\usage{
plot_synergy_interaction_index(ds, sample_type, treatment_1, treatment_2,
  proportion_1, hour, ..., log = "y", alpha = 0.05)
}
\arguments{
\item{ds}{dataset}

\item{sample_type}{sample type in ds}

\item{treatment_1}{treatment in ds}

\item{treatment_2}{treatment in ds}

\item{proportion_1}{proportion of treatment 1 in total concentration of 
synergy experiment}

\item{hour}{hour in ds. Default 0.}

\item{...}{additional parameters to pass to plot function}

\item{log}{which scale should be logged; default is "y"}

\item{alpha}{1-alpha is the size of the confidence intervals, default 0.05}
}
\value{
list ii: estimated interaction indices corresponding to the 
observations (c.d1, c.d2, E); ii.low, ii.up: estimated lower and upper CI
}
\description{
Calls code published by Lee and Kong in Statistics in Biopharmaceutical 
Research 2012; please cite their work if you use this function.
}
\examples{
dose_SCH=c(0.1, 0.5, 1, 2, 4)
eff_SCH = c(0.6701, 0.6289, 0.5577, 0.4550, 0.3755)
dose_4HPR = c(0.1, 0.5, 1, 2)
eff_4HPR = c(0.7666, 0.5833, 0.5706, 0.4934)
eff_comb = c(0.6539, 0.4919, 0.3551, 0.2341)
syn = data.frame( treatment_1 = rep("SCH66336", 13),
                  conc_1 = c( dose_SCH, rep(0, 4), dose_SCH[1:4]),
                  treatment_2 = rep("4-HPR", 13),
                  conc_2 = c( rep(0, 5), dose_4HPR, dose_4HPR ),
                  values = c(eff_SCH, eff_4HPR, eff_comb ) )
ds_lk = create_synergy_dataset( sample_types = rep("sample_1", 13), 
                                treatments_1 = syn$treatment_1,
                                treatments_2 = syn$treatment_2,
                                concentrations_1 = syn$conc_1,
                                concentrations_2 = syn$conc_2,
                                values = syn$values)
ii=plot_synergy_interaction_index(ds=ds_lk, sample_type = "sample_1", 
                              treatment_1 = "SCH66336", treatment_2="4-HPR", 
                              proportion_1=0.5, hour=0, 
                              xlab="Effect", ylab="Interaction Index" ,
                              main="Interaction Index")
}
\references{
Lee & Kong Statistics in Biopharmaceutical Research 2012
}

