% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timecourse_raw}
\alias{plot_timecourse_raw}
\title{Plot an time course of the raw intensities}
\usage{
plot_timecourse_raw(D, sample_types, treatments, concentrations, ...,
  cex.yaxis = 2, axis.font = 2, summary_method = "mean")
}
\arguments{
\item{D}{experiment dataset with columns matching the output of 
\code{combine_data_and_maps}}

\item{sample_types}{which sample types to draw}

\item{treatments}{which treatments to draw}

\item{concentrations}{which concentrations to draw}

\item{...}{standard parameters for \code{plot} function}

\item{cex.yaxis}{size coefficient for Y axis, default 2}

\item{axis.font}{font value for axis, default 2 (bold)}

\item{summary_method}{summary method for points to plot in timecourse, one 
of (mean, median), defaults to mean}
}
\value{
data frame reporting points plotted, sample types, and concentrations
}
\description{
Plot an time course of the raw intensities
}
\examples{
pkg = "HTDoseResponseCurve"
fn_map = system.file("extdata", "sample_data_384_platemap.txt",package=pkg)
fn_data = system.file("extdata", "sample_data_384.xlsx", package = pkg)
plate_map = read_platemap_from_Incucyte_XML( fn_map )
plate_data = read_plates_from_Incucyte_export( fn_data, "p1", 
                                               number_of_wells=384)
plate_data$hours = round(plate_data$hours)
ds = combine_data_and_map( plate_data, plate_map, negative_control = "DMSO" )
ds = normalize_plates_by_vehicle( ds, summary_method="mean")
ds = ds[ds$treatment=="drug13" | ds$treatment=="DMSO",]
plot_timecourse_raw( ds, sample_types=c("line_1", "line_2"), 
                     treatments="DMSO", concentrations=0)
}

