% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_values_by_plate}
\alias{plot_values_by_plate}
\title{Plot an image of the raw intensities for a plate}
\usage{
plot_values_by_plate(plate, hour, color_bounds = c(0, 100),
  color_palatte = c("white", "#fdbb84", "#e34a33"), main = "", cex = 1.5)
}
\arguments{
\item{plate}{data frame in format matching that produced by 
\code{read_incucyte_from_excel}.}

\item{hour}{hour of the experiment to plot}

\item{color_bounds}{values at which to draw coldest and hottest colors, 
defaults to c(0, 100)}

\item{color_palatte}{colors to use for cold-middle-hot, defaults to 
c("white", "#fdbb84","#e34a33")}

\item{main}{title to draw above plot, defaults to ""}

\item{cex}{point size, defaults to 1.5}
}
\value{
none
}
\description{
Plot an image of the raw intensities for a plate
}
\examples{
# small toy example
plate_1 = create_empty_plate( 6, hour=0, plate_id="plate_1")
plate_1[1,1:3] = c(100, 40, 10 )
plate_1[2,1:3] = c(90, 70, 30 )
plot_values_by_plate( plate_1, hour=0 )
# real-world example
pkg = "HTDoseResponseCurve"
fn_data = system.file("extdata", "sample_data_384.xlsx", package = pkg)
plate_data = read_plates_from_Incucyte_export( fn_data, "p1", 
                                               number_of_wells=384)
plot_values_by_plate(plate_data, hour=96)
}

