% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_dataset}
\alias{read_dataset}
\title{Read a dataset from a text file}
\usage{
read_dataset(filepath, negative_control = NA, plate_id = "plate_1",
  sep = "\\t")
}
\arguments{
\item{filepath}{Path to the file to load. If the full path is not given, R 
will assume the file is in the current working directory.}

\item{negative_control}{A designation for the negative controls in this 
dataset, if they exist. Value may be NA, a number, a string, or a data frame.

\itemize{
 \item{NA: Use when there are no negative control measurements. The contents 
   of the column named 'value_normalized' will be copied from the contents of 
   the column named 'value'. }
 \item{Number: Use when each treatment has been labeled with a concentration 
   (typically 0) that indicates the vehicle control. Each treatment must 
   contain one or more observations with this concentration, and these 
   observations will be the negative controls.}
 \item{string: Use when a single set of observations is a universal control. 
   The treatment whose name matches the string is the universal 
   negative control all of the data.}
 \item{data frame: Use when more than one negative control exists, and you 
   have to map different treatments to a particular negative control. The 
   data frame must have names 'drug' and 'vehicle', and the data frame will 
   map match treatments in the 'drug' column to those in the 
   'vehicle' column.}
}}

\item{plate_id}{A string identifying this set of observations. Defaults to 
"plate_1".}

\item{sep}{A character that separates elements in the text file. Fefaults to 
tab.}
}
\value{
A data frame for the dataset matching the output of the 
\code{\link{create_dataset}} function.
}
\description{
The text file may be separated by commas, tabs, or any other commonly used 
delimiter. The first row must have the following elements: 
\itemize{
 \item{sample_type}
 \item{treatment}
 \item{concentration}
 \item{value}
}
}
\details{
The text file may optionally contain a column called "hour", indicating the 
time point of the measurement. If "hour" is present the value must be a 
number.
}
\examples{
pkg = "HTDoseResponseCurve"
fn_txt = system.file("extdata", "sample_data_1.txt", package = pkg)
ds = read_dataset( fn_txt, negative_control="DMSO" )
}
\seealso{
\code{\link{create_dataset}}
}

