% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_platemap_from_excel}
\alias{read_platemap_from_excel}
\title{Read plate maps into a list of three data frames}
\usage{
read_platemap_from_excel(filename, sheet_num = 1, number_of_wells,
  concentration_identifier = "concentration",
  treatment_identifier = "treatment", sample_identifier = "cell line")
}
\arguments{
\item{filename}{Excel file generated by the user}

\item{sheet_num}{index of the Excel sheet to read from filename, defaults 1}

\item{number_of_wells}{number of wells in each plate}

\item{concentration_identifier}{text in a cell in column 1 that signals the 
concentration map will follow, defaults "concentration"}

\item{treatment_identifier}{text in a cell in column 1 that signals the 
treatment_identifier map will follow, defaults "treatment"}

\item{sample_identifier}{text in a cell in column 1 that signals the 
sample_identifier map will follow, defaults "cell line"}
}
\value{
a data frame where columns are data, timestamp, plate_id, hour
}
\description{
A Plate map describes the cell lines, treatment, and treatment concentrations
applied to each well. As an alternate to exporting a plate map into 
XML directly from the instrument, \code{read_platemap_from_excel} loads plate
map data from an Excel file. The input file must contain cells in column A
indicating the values for treatment concentration in uM (e.g. 0.2 for 200 nM), 
treatment (e.g. the drug in a well), and sample type (e.g. the identifier for 
a particular cell line in a well). These cells should be 
one cell up and to the left of plate maps themselves.
}

