% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_platemap_from_Incucyte_XML}
\alias{read_platemap_from_Incucyte_XML}
\title{Read plate maps from Incucyte plate map XML extract}
\usage{
read_platemap_from_Incucyte_XML(path_to_file, attribute_value = "description")
}
\arguments{
\item{path_to_file}{plate map file generated by the user}

\item{attribute_value}{The XML attribute to extract for treatment and sample 
names, one of {"description", "displayName"}, defaults to "description"}
}
\value{
a list of matrixes with the same dimension as the plate map. The 
matrixes are concentration, treatment, sample_type, seeding_density, and 
passage
}
\description{
A Plate map describes the cell lines, treatment, and treatment concentrations
applied to each well. This file is a text file in the XML format.
}

