% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_plates_from_Incucyte_export}
\alias{read_plates_from_Incucyte_export}
\title{Read Incucyte-formatted Excel document for a plate at one or more time points}
\usage{
read_plates_from_Incucyte_export(path_to_file, plate_id, number_of_wells,
  sheet_num = 1, plate_mask = NA)
}
\arguments{
\item{path_to_file}{Path to Excel file generated by the Incucyte software}

\item{plate_id}{a user-defined string to identify this plate. A plate can be 
imaged at more than one time point.}

\item{number_of_wells}{the number of wells per plate, must be 384 or 96}

\item{sheet_num}{the index of the Excel sheet to read, defaults 1}

\item{plate_mask}{determine which wells to read; if NA (default), read all 
wells. If a matrix of boolean values is passed, read only the wells where 
plate_mask is TRUE. The matrix must have the same dimensions as the wells 
being read.}
}
\value{
a data frame where columns are data, plate_id, hour.
}
\description{
\code{read_plates_from_Incucyte_export} loads data that have been exported to 
Excel from the Incucyte platform and returns a data frame of the results, 
with additional columns for the timestamp, hour, and a user-defined plate 
identifier. Each call to this function will load the exported values from one 
or more time points for a single plate map. Individual plate maps can be 
distinguished in later analysis by the user-provided plate_id parameter.
}
\details{
The number of columns will be the columns in the plate plus three 
additional columns: timestamp, plate_id, and hours. The number of rows will 
be the number of rows in the plate times the number of plates in the Excel 
sheet. Excel sheets are read using Hadley Wickham's readxl package.

If the Incucyte instrument only exports a partial plate (e.g. if it does not 
export every row) this function will including padding for rows and columns 
that are not in the extract.
}

