% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{summary.HT_fit}
\alias{summary.HT_fit}
\title{Summarize the results of a dose response fit}
\usage{
\method{summary}{HT_fit}(object, ...)
}
\arguments{
\item{object}{fit object}

\item{...}{other parameters, currently ignored}
}
\value{
none
}
\description{
Summarize the results of a dose response fit
}
\examples{
sample_types = rep( c(rep("line1",3), rep("line2",3)), 5)
treatments = c(rep("DMSO",6), rep("drug",24))
concentrations = c( rep(0,6),rep(200,6), rep(500,6),rep(1000,6),rep(5000,6))
values=c(100,99,100,90,91,92,99,97,99,89,87,88,86,89,88,56,59,58,66,65,67,
         25,23,24,42,43,46,4,5,9)
hours = rep(48, length(values))
plate_id = "plate_1"
ds = create_dataset( sample_types, treatments, concentrations, 
                      hours, values, plate_id, negative_control = "DMSO")
ds = normalize_plates_by_vehicle(ds, summary_method = "mean")
library(drc)
# Fit model using three-parameter log-logistic function
fit_1 = fit_DRC(ds, sample_types=c("line1", "line2"), treatments=c("drug"), 
                hour = 48, fct=drc::LL.3() )
summary(fit_1)
}

