% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_platemap_to_XML}
\alias{write_platemap_to_XML}
\title{Write a loaded plate map object in XML format}
\usage{
write_platemap_to_XML(fn, plate_map, passage, seeding_density, density_units,
  concentration_units)
}
\arguments{
\item{fn}{complete path of file to write}

\item{plate_map}{list with matrixes labeled concentration, treatment, and 
sample_type}

\item{passage}{integer number for passage}

\item{seeding_density}{integer number for seeding density}

\item{density_units}{string indicating the density units}

\item{concentration_units}{string indicating the treatment concentration 
units}
}
\value{
none
}
\description{
Write a loaded plate map object in XML format
}

