% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{combine_data_and_map}
\alias{combine_data_and_map}
\title{Combine raw data and plate map to produce a single tall data frame}
\usage{
combine_data_and_map(raw_plate, plate_map, negative_control = 0,
  summary_method = "mean")
}
\arguments{
\item{raw_plate}{data frame with format matching that produced by a call to 
\code{read_incucyte_exported_to_excel}}

\item{plate_map}{list where each item specifies concentrations, treatments, 
and cell lines for a plate specified in raw_plates, matching the format of a 
file created by \code{read_platemap_from_excel}}

\item{negative_control}{A designation for the negative controls in this 
dataset, if they exist. Value may be NA, a number, a string, or a data frame.

\itemize{
 \item{NA: Use when there are no negative control measurements. The contents 
   of the column named 'value_normalized' will be copied from the contents of 
   the column named 'value'. }
 \item{Number: Use when each treatment has been labeled with a concentration 
   (typically 0) that indicates the vehicle control. Each treatment must 
   contain one or more observations with this concentration, and these 
   observations will be the negative controls.}
 \item{string: Use when a single set of observations is a universal control. 
   The treatment whose name matches the string is the universal 
   negative control all of the data.}
 \item{data frame: Use when more than one negative control exists, and you 
   have to map different treatments to a particular negative control. The 
   data frame must have names 'drug' and 'vehicle', and the data frame will 
   map match treatments in the 'drug' column to those in the 
   'vehicle' column.}
}}

\item{summary_method}{summarize replicate measures by either mean or median; 
must be one of "mean", "median". Defaults to "mean"}
}
\value{
a data frame where columns indicate the sample type, treatment, 
concentration, observed raw value, normalized value, name of the 
negative_control treatment, whether a particular row 
is a negative control for at least one other row, hours since the start time,
and plate of origin
}
\description{
Combine a single data plate matching the output of 
\code{\link{create_empty_plate}} with a plate map matching the output of 
\code{\link{create_empty_plate_map}} to create a single tall data frame with 
one row per observation.
}
\details{
The user specifies which wells (if any) bear negative controls (i.e. 
vehicles) 
Negative controls:

If the experiment lacks negative controls, pass an empty string to the 
negative_control parameter.

If each treatment has wells with a concentration in the plate map that 
indicates the negative control for that treatment, pass that value to the 
negative_control parameter. The default value is 0.

If more all wells share a common negative control (e.g. DMSO), pass the name 
of the negative control treatment to the negative_control parameter.

If there is more than one distinct common negative control (e.g. DMSO for 
some treatments and ethanol for others), pass a data frame with column names 
"drug" and "vehicle" to indicate the appropriate vehicle for each drug.
If a data frame is passed, the drug column should contain an entry for each 
treatment on the plate(s).
}
\examples{
# Create a six well plate testing two drugs on two cell lines. Each drug has 
# two concentrations plus a DMSO well. Plan to normalize against DMSO.
# Obviously real data would include replicates, more concentrations, etc.
plate = create_empty_plate( 6, hour=0, plate_id="plate_1")
plate[1:2,1:3] = c(99,98,90,87,77,20)
plate_map = create_empty_plate_map( number_of_wells = 6 )
plate_map$concentration[1:2,1:3] = c( 0, 0, 100, 100, 200, 200 )
plate_map$treatment[1:2,1:3] = c("DMSO","DMSO","drug1","drug2","drug1",
                                 "drug2")
plate_map$sample_type[1:2,1:3] = rep( c("line1", "line2"), 3)
combine_data_and_map( plate, plate_map, negative_control="DMSO" )

# Create a six well plate testing two drugs on two cell lines. Each drug has 
# three concentrations, one of which is vehicle-only and labeled zero.
plate = create_empty_plate( 6, hour=0, plate_id="plate_1")
plate[1:2,1:3] = c(99,98,90,87,77,20)
plate_map = create_empty_plate_map( number_of_wells = 6 )
plate_map$concentration[1:2,1:3] = c( 0, 0, 100, 100, 200, 200 )
plate_map$treatment[1:2,1:3] = c("drug1","drug2","drug1","drug2","drug1",
                                 "drug2")
plate_map$sample_type[1:2,1:3] = rep( c("line1", "line2"), 3)
combine_data_and_map( plate, plate_map, negative_control=0 )
}

