% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{create_empty_plate}
\alias{create_empty_plate}
\title{Create a plate list with empty elements}
\usage{
create_empty_plate(number_of_wells, hour = 0, plate_id = "plate_1")
}
\arguments{
\item{number_of_wells}{Number of wells in the plate, must be one of 
{6, 12, 24, 96, 384}.}

\item{hour}{Number of hours since the experiment began. Defaults to 0.}

\item{plate_id}{Text string that identifies this plate map. Defaults to 
"plate_1".}
}
\value{
A data frame
}
\description{
Create a plate list with empty elements
}
\examples{
create_empty_plate( number_of_wells=6 )
create_empty_plate( number_of_wells=96, hour=0, plate_id="plate_12")
create_empty_plate( number_of_wells=384, hour=12, plate_id="plate_9")
}
\seealso{
\code{\link{create_empty_plate_map}}
}

