% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{create_empty_plate_map}
\alias{create_empty_plate_map}
\title{Create an empty plate map}
\usage{
create_empty_plate_map(number_of_wells, max_treatments_per_well = 1)
}
\arguments{
\item{number_of_wells}{Number of wells in the plate, must be one of 
{6, 12, 24, 96, 384}}

\item{max_treatments_per_well}{The maximum number of treatment 
conditions per each well, either 1 or 2. Defaults to 1. Pass 2 only if 
performing a synergy experiment.}
}
\value{
a plate_map list with matrixes for treatment, concentration, 
sample_types, density, and passage
}
\description{
The plate map object is a list of identically sized matrixes named 
treatment, concentration, sample_type, density, and passage. By default, only 
a single concentration and treatment is expected each individual well. If 
this is a synergy experiment and each well can contain two or more different 
treatments at varying concentrations, pass the maximum number of conditions 
per well to conditions_per_well. This will create additional matrixes, 
e.g. treatment_2 and concentration_2.
}
\examples{
create_empty_plate_map( number_of_wells=96 )
}
\seealso{
\code{\link{create_empty_plate}}
}

