% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTDoseResponseCurve.R
\name{get_concentrations}
\alias{get_concentrations}
\title{Sorted unique list of treatment concentrations}
\usage{
get_concentrations(D)
}
\arguments{
\item{D}{A plate map list or a data frame generated by 
\code{combine_data_and_map}}
}
\value{
sorted vector of concentrations, including 0 if present
}
\description{
Convenience method to extract treatment concentrations
}
\examples{
pm = create_empty_plate_map( 6 )
pm$concentration[1,1:3] = c( 0, 100, 200 )
pm$concentration[2,1:3] = c( 0, 100, 300 )
get_concentrations( pm )
ds = create_dataset( 
  sample_types= c("line1","line1","line1","line2","line2","line2"),
  treatments = c("DMSO","drug1","drug2","DMSO","drug1","drug2"),
  concentrations = c(0, 100, 200, 0, 100, 200),
  hours = c(48, 48, 48, 48, 48, 48),
  values = c(100, 90, 20, 100, 89, 87), 
  plate_id = "plate_1",
  negative_control = "DMSO")
get_concentrations( ds )
}
\seealso{
\code{\link{get_sample_types}}

\code{\link{get_treatments}}

\code{\link{get_hours}}
}

