% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_platemap_from_Incucyte_XML}
\alias{read_platemap_from_Incucyte_XML}
\title{Read plate maps from Incucyte plate map XML extract}
\usage{
read_platemap_from_Incucyte_XML(path_to_file, attribute_value = "description",
  max_treatments_per_well = 1)
}
\arguments{
\item{path_to_file}{The complete path to a plate map file.}

\item{attribute_value}{The XML attribute to extract for treatment and sample 
names, one of {"description", "displayName"}. Defaults to "description".}

\item{max_treatments_per_well}{A number indicating the maximum number of 
treatments to expect in a single well. For single-treatment experiments, pass 
1. For synergy experiments where two treatments may be combined in a single 
well, pass 2. Defaults to 1.}
}
\value{
A list of matrixes with the same dimension as the plate map. The 
matrixes are concentration, treatment, sample_type, seeding_density, and 
passage
}
\description{
A Plate map describes the cell lines, treatment, and treatment concentrations
applied to each well. This file is a text file in the XML format.
}
\examples{
pkg = "HTDoseResponseCurve"
fn_map = system.file("extdata", "sample_data_384_platemap.txt",package=pkg)
plate_map = read_platemap_from_Incucyte_XML( fn_map )
}

