% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy.R
\name{synergy_interaction_CI}
\alias{synergy_interaction_CI}
\title{Construct confidence intervals for observed effects at combination doses 
having observed effects.}
\usage{
synergy_interaction_CI(ds, sample_type, treatment_1, treatment_2, hour = 0,
  alpha = 0.05, summary_method = "mean")
}
\arguments{
\item{ds}{dataset}

\item{sample_type}{sample type in ds}

\item{treatment_1}{treatment in ds}

\item{treatment_2}{treatment in ds}

\item{hour}{hour in ds. Default 0.}

\item{alpha}{1-alpha is the size of the confidence intervals, default 0.05}

\item{summary_method}{mean or median}
}
\value{
A list with the interaction index, lower, and upper confidence 
intervals for fix effects points in E
}
\description{
If the dataset does not have negative controls, measurements at 
concentration zero are assumed to be empty for a given treatment.
Adapted directly from code published in 
Lee and Kong Statistics in Biopharmaceutical Research 2012
}
\references{
Lee & Kong Statistics in Biopharmaceutical Research 2012
}

