% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_platemap_to_XML}
\alias{write_platemap_to_XML}
\title{Write a loaded plate map object in XML format}
\usage{
write_platemap_to_XML(fn, plate_map, passage, seeding_density, density_units,
  concentration_units)
}
\arguments{
\item{fn}{A complete path to the file to write.}

\item{plate_map}{A list with matrixes labeled concentration, treatment, and 
sample_type.}

\item{passage}{Number for cell passage.}

\item{seeding_density}{Number for seeding density.}

\item{density_units}{Text string indicating the density units.}

\item{concentration_units}{Text string indicating the treatment concentration 
units.}
}
\value{
none
}
\description{
Write a loaded plate map object in XML format
}
\examples{
pkg = "HTDoseResponseCurve"
fn_map = system.file("extdata", "sample_data_96_platemap.xlsx",package=pkg)
plate_map = read_platemap_from_excel( fn_map, number_of_wells=96 )
#write_platemap_to_XML( fn_map, plate_map, passage=1, seeding_density=1000,
#  density_units="thousands", concentration_units="ng/ul")
}

